LayoutTests/fast/dom/elementFromPoint-relative-to-viewport-expected.txt

This test document.elementFromPoint is evaluated in with respect to the viewport, not the document.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS unscrolledBoxInitial is '0'
PASS scrolledDownBoxInitial is '15'
PASS scrolledRightBoxInitial is '3'
PASS scrolledDownAndRightBoxInitial is '18'
PASS unscrolledBoxOffscreen is '0'
PASS scrolledDownBoxOffscreen is '15'
PASS scrolledRightBoxOffscreen is '3'
PASS scrolledDownAndRightBoxOffscreen is '18'
PASS successfullyParsed is true

TEST COMPLETE
